requirements.txt
numpy
scipy
torch
faiss-cpu  # or faiss-gpu
tqdm
scikit-learn